move `ExtensionTipsService`
