+++ src/vs/platform/extensionManagement/node/extensionTipsService.ts
